requirements.txt
google-generativeai
python-telegram-bot
pydantic
schedule
python-dotenv
chromadb
duckduckgo-search
psutil
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Testing
pytest
pytest-asyncio
pytest-mock
